WIP
Move existing benchmarks to WasmLegacyBenchamrk and add a new WasmBenchmark that supports
running iterations. It's currently an async benchmark subclass since the compile is async.

As a proof of concept this patch converts tsf-wasm to use the new scoring. As part of this
process tsf needed to be recompiled so I added a step to the build.sh script which wraps
emcc's output .js file with some glue code. This should make it easier to rebuild in the
future too since we're not directly editing the module loading file. I went with this
route rather than hooking into emcc's API since I wanted to use the js glue as close to
"out of the box" as possible.

Lastly, get emcc to ouput the .symbols object so we know which wasm function is which index.
We could emit a name section into he binary but my assumption is that most production code
does not ship with that so it could change the benchmark in an unrealistic way.

TODO:
* Profile perf differences
* Generalize the wrapping process
* Remove postIterationHook code for a different PR.