Move quicksort-wasm to JS-style scoring
The usual changes: use `WasmEMCCBenchmark` class for running, rebuild and add build script, use iteration count that makes the Wall time not too high (<2s)

In this case, we spend a bit more time in the top-tier than before. However, that line item earlier ran a single <50ms iteration, and compiled for ~2ms, so I would argue it was very prone to noise, and it's good that we run it for longer.