Move prepareForNextIteration call from Benchmark's runnerCode to preIterationCode
This makes it easier for Benchmark subclasses to also call. This fixes a bug in https://github.com/WebKit/JetStream/pull/74
which caused inspector code load benchmarks to not set up properly ending the benchmark in 0ms.

Also, use optional chaining for the call (and few other similar optional helpers).