Configurable resource prefetching  (#70)
This addresses issue #58 and is useful for local profiling.

If the prefetchResources is disabled, we now no longer preload them eagerly, but rather use the live URL and either script-inject it or use it directly with workers.
This means we can easily get the original source positions for profiling.

- `--no-prefetch` command ilne support
- `&prefetchResources=0` URLParam support
- Improve test coverage and add --no-prefetch tests
- Check that test shell output for failures to catch JSC errors (it does not report exit != 0 on rejected promises)

Drive-by-fix:
- Move FileLoader out of the IIFE
- Check that JetStreamDriver has at least one active benchmark
- Add assertions for valid scores
- Fix benchmarks selection with tags