Dotnet Uses Legacy Wasm Exceptions.
Add a pass to the build script that replaces them with the non-legacy
execptions. This doesn't seem to impact performance. The dotnet runtime
still seems to generate one module/function that has a legacy exception
but that function is never called. It's also somewhat degenerate
since it's just `(try catch_all)` and has no other body.

This doesn't seem to make a difference for the score in all CLIs and we
don't want legacy exception usage in JetStream to be the thing that
keeps the legacy wasm exceptions "feature" alive.

Note: I didn't run the build script. I just ran the wasm-opt
commands on the build products. Mostly because when I recompile
the code doesn't run or is 5x slower than the previous binaries.